requirements.txt
pyrofork
motor
tgcrypto
py-tgcalls
ntgcalls
aiohttp
pymongo
dnspython
Pillow
python-dotenv
youtube-search-python